retrieval-service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
faiss-cpu>=1.8.0
numpy>=1.24.3
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
python-docx>=0.8.11
